Updates per code review comments. Renamed platform.browser to platform.dom and move associated tests into platform.dom.tests. Chart.platform is now a static class instead of a constructable object.
